+++ lib/src/service/iso_alpha2_country.dart
@@ -1,0 +1,751 @@
+// Copyright 2022 Kato Shinya. All rights reserved.
+// Redistribution and use in source and binary forms, with or without
+// modification, are permitted provided the conditions.
+
+enum ISOAlpha2Country {
+  /// `AF`
+  afghanistan('AF'),
+
+  /// `AX`
+  alandIslands('AX'),
+
+  /// `AL`
+  albania('AL'),
+
+  /// `DZ`
+  algeria('DZ'),
+
+  /// `AS`
+  americanSamoa('AS'),
+
+  /// `AD`
+  andorra('AD'),
+
+  /// `AO`
+  angola('AO'),
+
+  /// `AI`
+  anguilla('AI'),
+
+  /// `AQ`
+  antarctica('AQ'),
+
+  /// `AG`
+  antiguaAndBarbuda('AG'),
+
+  /// `AR`
+  argentina('AR'),
+
+  /// `AM`
+  armenia('AM'),
+
+  /// `AW`
+  aruba('AW'),
+
+  /// `AU`
+  australia('AU'),
+
+  /// `AT`
+  austria('AT'),
+
+  /// `AZ`
+  azerbaijan('AZ'),
+
+  /// `BS`
+  bahamas('BS'),
+
+  /// `BH`
+  bahrain('BH'),
+
+  /// `BD`
+  bangladesh('BD'),
+
+  /// `BB`
+  barbados('BB'),
+
+  /// `BY`
+  belarus('BY'),
+
+  /// `BE`
+  belgium('BE'),
+
+  /// `BZ`
+  belize('BZ'),
+
+  /// `BJ`
+  benin('BJ'),
+
+  /// `BM`
+  bermuda('BM'),
+
+  /// `BT`
+  bhutan('BT'),
+
+  /// `BO`
+  bolivia('BO'),
+
+  /// `BQ`
+  bonaire('BQ'),
+
+  /// `BA`
+  bosniaAndHerzegovina('BA'),
+
+  /// `BW`
+  botswana('BW'),
+
+  /// `BV`
+  bouvetIsland('BV'),
+
+  /// `BR`
+  brazil('BR'),
+
+  /// `IO`
+  britishIndianOceanTerritory('IO'),
+
+  /// `BN`
+  bruneiDarussalam('BN'),
+
+  /// `BG`
+  bulgaria('BG'),
+
+  /// `BF`
+  burkinaFaso('BF'),
+
+  /// `BI`
+  burundi('BI'),
+
+  /// `KH`
+  cambodia('KH'),
+
+  /// `CM`
+  cameroon('CM'),
+
+  /// `CA`
+  canada('CA'),
+
+  /// `CV`
+  capeVerde('CV'),
+
+  /// `KY`
+  caymanIslands('KY'),
+
+  /// `CF`
+  centralAfricanRepublic('CF'),
+
+  /// `TD`
+  chad('TD'),
+
+  /// `CL`
+  chile('CL'),
+
+  /// `CN`
+  china('CN'),
+
+  /// `CX`
+  christmasIsland('CX'),
+
+  /// `CC`
+  cocosIslands('CC'),
+
+  /// `CO`
+  colombia('CO'),
+
+  /// `KM`
+  comoros('KM'),
+
+  /// `CD`
+  congo('CD'),
+
+  /// `CK`
+  cookIslands('CK'),
+
+  /// `CR`
+  costaRica('CR'),
+
+  /// `CI`
+  ivoryCoast('CI'),
+
+  /// `HR`
+  croatia('HR'),
+
+  /// `CU`
+  cuba('CU'),
+
+  /// `CW`
+  curacao('CW'),
+
+  /// `CY`
+  cyprus('CY'),
+
+  /// `CZ`
+  czechRepublic('CZ'),
+
+  /// `DK`
+  denmark('DK'),
+
+  /// `DJ`
+  djibouti('DJ'),
+
+  /// `DM`
+  dominica('DM'),
+
+  /// `DO`
+  dominicanRepublic('DO'),
+
+  /// `EC`
+  ecuador('EC'),
+
+  /// `EG`
+  egypt('EG'),
+
+  /// `SV`
+  elSalvador('SV'),
+
+  /// `GQ`
+  equatorialGuinea('GQ'),
+
+  /// `ER`
+  eritrea('ER'),
+
+  /// `EE`
+  estonia('EE'),
+
+  /// `ET`
+  ethiopia('ET'),
+
+  /// `FK`
+  falklandIslands('FK'),
+
+  /// `FO`
+  faroeIslands('FO'),
+
+  /// `FJ`
+  fiji('FJ'),
+
+  /// `FI`
+  finland('FI'),
+
+  /// `FR`
+  france('FR'),
+
+  /// `GF`
+  frenchGuiana('GF'),
+
+  /// `PF`
+  frenchPolynesia('PF'),
+
+  /// `TF`
+  frenchSouthernTerritories('TF'),
+
+  /// `GA`
+  gabon('GA'),
+
+  /// `GM`
+  gambia('GM'),
+
+  /// `GE`
+  georgia('GE'),
+
+  /// `DE`
+  germany('DE'),
+
+  /// `GH`
+  ghana('GH'),
+
+  /// `GI`
+  gibraltar('GI'),
+
+  /// `GR`
+  greece('GR'),
+
+  /// `GL`
+  greenland('GL'),
+
+  /// `GD`
+  grenada('GD'),
+
+  /// `GP`
+  guadeloupe('GP'),
+
+  /// `GU`
+  guam('GU'),
+
+  /// `GT`
+  guatemala('GT'),
+
+  /// `GG`
+  guernsey('GG'),
+
+  /// `GN`
+  guinea('GN'),
+
+  /// `GW`
+  guineaBissau('GW'),
+
+  /// `GY`
+  guyana('GY'),
+
+  /// `HT`
+  haiti('HT'),
+
+  /// `HM`
+  heardIslandAndMcDonaldIslands('HM'),
+
+  /// `VA`
+  holySee('VA'),
+
+  /// `HN`
+  honduras('HN'),
+
+  /// `HK`
+  hongKong('HK'),
+
+  /// `HU`
+  hungary('HU'),
+
+  /// `IS`
+  iceland('IS'),
+
+  /// `IN`
+  india('IN'),
+
+  /// `ID`
+  indonesia('ID'),
+
+  /// `IR`
+  iran('IR'),
+
+  /// `IQ`
+  iraq('IQ'),
+
+  /// `IE`
+  ireland('IE'),
+
+  /// `IM`
+  isleOfMan('IM'),
+
+  /// `IL`
+  israel('IL'),
+
+  /// `IT`
+  italy('IT'),
+
+  /// `JM`
+  jamaica('JM'),
+
+  /// `JP`
+  japan('JP'),
+
+  /// `JE`
+  jersey('JE'),
+
+  /// `JO`
+  jordan('JO'),
+
+  /// `KZ`
+  kazakhstan('KZ'),
+
+  /// `KE`
+  kenya('KE'),
+
+  /// `KI`
+  kiribati('KI'),
+
+  /// `KR`
+  korea('KR'),
+
+  /// `KW`
+  kuwait('KW'),
+
+  /// `KG`
+  kyrgyzstan('KG'),
+
+  /// `LA`
+  laoPeoplesDemocraticRepublic('LA'),
+
+  /// `LV`
+  latvia('LV'),
+
+  /// `LB`
+  lebanon('LB'),
+
+  /// `LS`
+  lesotho('LS'),
+
+  /// `LR`
+  liberia('LR'),
+
+  /// `LY`
+  libya('LY'),
+
+  /// `LI`
+  liechtenstein('LI'),
+
+  /// `LT`
+  lithuania('LT'),
+
+  /// `LU`
+  luxembourg('LU'),
+
+  /// `MO`
+  macao('MO'),
+
+  /// `MK`
+  macedonia('MK'),
+
+  /// `MG`
+  madagascar('MG'),
+
+  /// `MW`
+  malawi('MW'),
+
+  /// `MY`
+  malaysia('MY'),
+
+  /// `MV`
+  maldives('MV'),
+
+  /// `ML`
+  mali('ML'),
+
+  /// `MT`
+  malta('MT'),
+
+  /// `MH`
+  marshallIslands('MH'),
+
+  /// `MQ`
+  martinique('MQ'),
+
+  /// `MR`
+  mauritania('MR'),
+
+  /// `MU`
+  mauritius('MU'),
+
+  /// `YT`
+  mayotte('YT'),
+
+  /// `MX`
+  mexico('MX'),
+
+  /// `FM`
+  micronesia('FM'),
+
+  /// `MD`
+  moldova('MD'),
+
+  /// `MC`
+  monaco('MC'),
+
+  /// `MN`
+  mongolia('MN'),
+
+  /// `ME`
+  montenegro('ME'),
+
+  /// `MS`
+  montserrat('MS'),
+
+  /// `MA`
+  morocco('MA'),
+
+  /// `MZ`
+  mozambique('MZ'),
+
+  /// `MM`
+  myanmar('MM'),
+
+  /// `NA`
+  namibia('NA'),
+
+  /// `NR`
+  nauru('NR'),
+
+  /// `NP`
+  nepal('NP'),
+
+  /// `NL`
+  netherlands('NL'),
+
+  /// `NC`
+  newCaledonia('NC'),
+
+  /// `NZ`
+  newZealand('NZ'),
+
+  /// `NI`
+  nicaragua('NI'),
+
+  /// `NE`
+  niger('NE'),
+
+  /// `NG`
+  nigeria('NG'),
+
+  /// `NU`
+  niue('NU'),
+
+  /// `NF`
+  norfolkIsland('NF'),
+
+  /// `MP`
+  northernMarianaIslands('MP'),
+
+  /// `NO`
+  norway('NO'),
+
+  /// `OM`
+  oman('OM'),
+
+  /// `PK`
+  pakistan('PK'),
+
+  /// `PW`
+  palau('PW'),
+
+  /// `PS`
+  palestine('PS'),
+
+  /// `PA`
+  panama('PA'),
+
+  /// `PG`
+  papuaNewGuinea('PG'),
+
+  /// `PY`
+  paraguay('PY'),
+
+  /// `PE`
+  peru('PE'),
+
+  /// `PH`
+  philippines('PH'),
+
+  /// `PN`
+  pitcairn('PN'),
+
+  /// `PL`
+  poland('PL'),
+
+  /// `PT`
+  portugal('PT'),
+
+  /// `PR`
+  puertoRico('PR'),
+
+  /// `QA`
+  qatar('QA'),
+
+  /// `RE`
+  reunion('RE'),
+
+  /// `RO`
+  romania('RO'),
+
+  /// `RU`
+  russianFederation('RU'),
+
+  /// `RW`
+  rwanda('RW'),
+
+  /// `BL`
+  saintBarthelemy('BL'),
+
+  /// `SH`
+  saintHelena('SH'),
+
+  /// `KN`
+  saintKittsAndNevis('KN'),
+
+  /// `LC`
+  saintLucia('LC'),
+
+  /// `MF`
+  saintMartin('MF'),
+
+  /// `PM`
+  saintPierreAndMiquelon('PM'),
+
+  /// `VC`
+  saintVincentAndTheGrenadines('VC'),
+
+  /// `WS`
+  samoa('WS'),
+
+  /// `SM`
+  sanMarino('SM'),
+
+  /// `ST`
+  saoTomeAndPrincipe('ST'),
+
+  /// `SA`
+  saudiArabia('SA'),
+
+  /// `SN`
+  senegal('SN'),
+
+  /// `RS`
+  serbia('RS'),
+
+  /// `SC`
+  seychelles('SC'),
+
+  /// `SL`
+  sierraLeone('SL'),
+
+  /// `SG`
+  singapore('SG'),
+
+  /// `SX`
+  sintMaarten('SX'),
+
+  /// `SK`
+  slovakia('SK'),
+
+  /// `SI`
+  slovenia('SI'),
+
+  /// `SB`
+  solomonIslands('SB'),
+
+  /// `SO`
+  somalia('SO'),
+
+  /// `ZA`
+  southAfrica('ZA'),
+
+  /// `GS`
+  southGeorgiaAndTheSouthSandwichIslands('GS'),
+
+  /// `SS`
+  southSudan('SS'),
+
+  /// `ES`
+  spain('ES'),
+
+  /// `LK`
+  sriLanka('LK'),
+
+  /// `SD`
+  sudan('SD'),
+
+  /// `SR`
+  suriname('SR'),
+
+  /// `SJ`
+  svalbardAndJanMayen('SJ'),
+
+  /// `SZ`
+  swaziland('SZ'),
+
+  /// `SE`
+  sweden('SE'),
+
+  /// `CH`
+  switzerland('CH'),
+
+  /// `SY`
+  syrianArabRepublic('SY'),
+
+  /// `TW`
+  taiwan('TW'),
+
+  /// `TJ`
+  tajikistan('TJ'),
+
+  /// `TZ`
+  tanzania('TZ'),
+
+  /// `TH`
+  thailand('TH'),
+
+  /// `TL`
+  timorLeste('TL'),
+
+  /// `TG`
+  togo('TG'),
+
+  /// `TK`
+  tokelau('TK'),
+
+  /// `TO`
+  tonga('TO'),
+
+  /// `TT`
+  trinidadAndTobago('TT'),
+
+  /// `TN`
+  tunisia('TN'),
+
+  /// `TR`
+  turkey('TR'),
+
+  /// `TM`
+  turkmenistan('TM'),
+
+  /// `TC`
+  turksAndCaicosIslands('TC'),
+
+  /// `TV`
+  tuvalu('TV'),
+
+  /// `UG`
+  uganda('UG'),
+
+  /// `UA`
+  ukraine('UA'),
+
+  /// `AE`
+  unitedArabEmirates('AE'),
+
+  /// `GB`
+  unitedKingdom('GB'),
+
+  /// `US`
+  unitedStates('US'),
+
+  /// `UM`
+  unitedStatesMinorOutlyingIslands('UM'),
+
+  /// `UY`
+  uruguay('UY'),
+
+  /// `UZ`
+  uzbekistan('UZ'),
+
+  /// `VU`
+  vanuatu('VU'),
+
+  /// `VE`
+  venezuela('VE'),
+
+  /// `VN`
+  vietnam('VN'),
+
+  /// `VG`
+  virginIslandsBritish('VG'),
+
+  /// `VI`
+  virginIslandsUS('VI'),
+
+  /// `WF`
+  wallisAndFutuna('WF'),
+
+  /// `EH`
+  westernSahara('EH'),
+
+  /// `YE`
+  yemen('YE'),
+
+  /// `ZM`
+  zambia('ZM'),
+
+  /// `ZW`
+  zimbabwe('ZW');
+
+  /// The code
+  final String code;
+
+  const ISOAlpha2Country(this.code);
+}